Project Euler Problem Generator
Still needs lots of work that I will have to finish later.

This is a WinForms application that will be used to generate folders for
each Problem Euler problem and will create a text file for each problem
if possible.
The texts files generated will include the problem's title and
description.